requirements.txt
robotframework==3.1.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.1.0
selenium==3.141.0
